refactor: split metrics into files - use MetricsNotificationsInterface::BlockConnected for (most) block metrics
