+++ src/metrics/container.h
@@ -2,8 +2,13 @@
 #define BITCOIN_METRICS_CONTAINER_H
 
 #include <cassert>
+#include <metrics/block.h>
+#include <metrics/mempool.h>
 #include <metrics/metrics.h>
+#include <metrics/net.h>
+#include <metrics/peer.h>
 #include <metrics/rpc.h>
+#include <metrics/tx.h>
 
 namespace metrics {
 static auto prom_registry = std::make_shared<prometheus::Registry>();
@@ -16,7 +21,6 @@ protected:
     std::unique_ptr<NetMetrics> _netMetrics;
     std::unique_ptr<TxMetrics> _txMetrics;
     std::unique_ptr<BlockMetrics> _blocks_metrics;
-    //std::unique_ptr<UtxoMetrics> _utxo_metrics;
     std::unique_ptr<MemPoolMetrics> _mempool_metrics;
     std::unique_ptr<ConfigMetrics> _cfg_metrics;
     std::unique_ptr<RpcMetrics> _rpc_metrics;
