+++ src/metrics/metrics.h
@@ -7,6 +7,7 @@
 #include <prometheus/histogram.h>
 #include <prometheus/registry.h>
 #include <prometheus/summary.h>
+
 namespace metrics {
 
 enum NetDirection {
@@ -133,243 +134,10 @@ public:
     void SetIBD(const bool value);
 };
 
-class BlockMetrics
-{
-public:
-    static std::unique_ptr<BlockMetrics> make(const std::string& chain, prometheus::Registry& registry, bool noop);
-    virtual ~BlockMetrics(){};
-    virtual void Size(size_t amt){};
-    virtual void SizeWitness(size_t amt){};
-    virtual void Height(int amt){};
-    virtual void Weight(size_t amt){};
-    virtual void Version(double amt){};
-    virtual void Transactions(size_t amt){};
-    virtual void SigOps(int64_t amt) {}
-    virtual void HeaderTime(int64_t amt) {}
-    virtual void Reward(int64_t amt){};
-    virtual void Fees(int64_t amt){};
-    virtual void Difficulty(double amt){};
-    virtual void ValueOut(double amt){};
 
-    virtual void TipLoadBlockDisk(int64_t current, double avg){};
-    virtual void TipConnectBlock(int64_t current, double avg){};
-    virtual void TipFlushView(int64_t current, double avg){};
-    virtual void TipFlushDisk(int64_t current, double avg){};
-    virtual void TipUpdate(int64_t current, double avg){};
 
-    virtual void ForkCheck(int64_t current, double avg){};
-    virtual void UpdateIndex(int64_t current, double avg){};
-};
 
-class BlockMetricsImpl : virtual public BlockMetrics, Metrics
-{
-protected:
-    void set(const std::string& type, double amt);
-    std::vector<std::string> _block_types{
-        "size",
-        "size-witness",
-        "weight",
-        "height",
-        "version",
-        "transactions",
-        "sigops",
-        "time",
-        "header-time",
-        "fees",
-        "reward",
-        "difficulty",
-        "valueout"};
-    std::map<const std::string, prometheus::Gauge*> _block_tip_gauge;
-    std::vector<prometheus::Histogram*> _block_bucket_timers;
-    std::vector<prometheus::Gauge*> _block_avg;
 
-public:
-    ~BlockMetricsImpl(){};
-    explicit BlockMetricsImpl(const std::string& chain, prometheus::Registry& registry);
-    void Size(size_t amt) override;
-    void SizeWitness(size_t amt) override;
-    void Height(int amt) override;
-    void Weight(size_t amt) override;
-    void Version(double amt) override;
-    void Transactions(size_t amt) override;
-    void SigOps(int64_t amt) override;
-    void HeaderTime(int64_t amt) override;
-    void Reward(int64_t amt) override;
-    void Fees(int64_t amt) override;
-    void Difficulty(double amt) override;
-    virtual void ValueOut(double amt) override;
-
-    void TipLoadBlockDisk(int64_t current, double avg) override;
-    void TipConnectBlock(int64_t current, double avg) override;
-    void TipFlushView(int64_t current, double avg) override;
-    void TipFlushDisk(int64_t current, double avg) override;
-    void TipUpdate(int64_t current, double avg) override;
-
-    void ForkCheck(int64_t current, double avg) override;
-    void UpdateIndex(int64_t current, double avg) override;
-};
-
-class TxMetrics
-{
-public:
-    virtual ~TxMetrics() {}
-    static std::unique_ptr<TxMetrics> make(const std::string& chain, prometheus::Registry& registry, bool noop);
-    virtual void IncInvalid(const std::string& reason){};
-    virtual void InputTime(double t){};
-    virtual void IncOrphanAdd(){};
-    virtual void IncOrphanRemove(){};
-    virtual void CacheSize(double amt){};
-    virtual void IncTransactions(const std::string& type, long amt){};
-    virtual void TransactionCheck(int64_t current, double amt){};
-};
-
-class TxMetricsImpl : virtual public TxMetrics, Metrics
-{
-protected:
-    std::map<const std::string, prometheus::Counter*> _single_transaction_counter;
-    prometheus::Summary* _inputs_timer;
-    prometheus::Counter* _orphan_remove_counter;
-    prometheus::Counter* _orphan_add_counter;
-    prometheus::Counter* _accepted_counter;
-    prometheus::Gauge* _cache_gauge;
-    prometheus::Histogram* _check_buckets;
-    prometheus::Gauge* _check_avg;
-
-public:
-    ~TxMetricsImpl() {}
-    explicit TxMetricsImpl(const std::string& chain, prometheus::Registry& registry);
-    void IncInvalid(const std::string& reason) override;
-    void InputTime(double t) override;
-    void IncOrphanAdd() override;
-    void IncOrphanRemove() override;
-    void CacheSize(double amt) override;
-    void IncTransactions(const std::string& type, long amt) override;
-    void TransactionCheck(int64_t current, double amt) override;
-};
-
-class NetMetrics
-{
-public:
-    virtual ~NetMetrics() {}
-    static std::unique_ptr<NetMetrics> make(const std::string& chain, prometheus::Registry& registry, bool noop);
-    virtual void IncConnection(const std::string& type){};
-    virtual void ConnectionGauge(NetConnectionType netConnection, uint amt){};
-    virtual void BandwidthGauge(NetDirection direction, const std::string& msg, uint64_t amt){};
-    virtual void PingTime(long amt){};
-    virtual void IncPingProblem(){};
-    virtual void MaxOutbound(int64_t amt){};
-    virtual void MaxOutboundStartTime(int64_t amt){};
-};
-class NetMetricsImpl : virtual public NetMetrics, Metrics
-{
-protected:
-    void initBandwidth();
-    std::map<NetConnectionType, prometheus::Gauge*> _connections_gauge;
-    std::map<const std::string, prometheus::Gauge*> _bandwidth_gauge_tx;
-    std::map<const std::string, prometheus::Gauge*> _bandwidth_gauge_rx;
-    prometheus::Summary* _ping_timer;
-    prometheus::Counter* _ping_problem_counter;
-    std::map<const std::string, prometheus::Counter*> _connection_counter;
-    prometheus::Gauge* _max_outbound_gauge;
-    prometheus::Gauge* _max_outbound_start_gauge;
-
-public:
-    ~NetMetricsImpl(){};
-    explicit NetMetricsImpl(const std::string& chain, prometheus::Registry& registry);
-    void IncConnection(const std::string& type) override;
-    void ConnectionGauge(NetConnectionType netConnection, uint amt) override;
-    void BandwidthGauge(NetDirection direction, const std::string& msg, uint64_t amt) override;
-    void PingTime(long amt) override;
-    void IncPingProblem() override;
-    void MaxOutbound(int64_t amt) override;
-    void MaxOutboundStartTime(int64_t amt) override;
-};
-
-class PeerMetrics
-{
-public:
-    virtual ~PeerMetrics() {}
-    static std::unique_ptr<PeerMetrics> make(const std::string& chain, prometheus::Registry& registry, bool noop);
-    virtual void ProcessMsgType(const std::string& msg_type, long amt){};
-    virtual void IncTxValidationResult(int state){};
-    virtual void IncDiscourage(){};
-    virtual void IncMisbehaveAmount(int amt){};
-    virtual void ConnectionType(int type, uint amt){};
-    virtual void Permission(const std::string& permission, uint amt){};
-    virtual void Known(size_t amt){};
-    virtual void SendMessageTime(long amt){};
-    virtual void PushMsgType(const std::string& msg_type){};
-    virtual void Banned(unsigned long amt){};
-};
-class PeerMetricsImpl : virtual public PeerMetrics, Metrics
-{
-protected:
-    void initConnections();
-    std::vector<prometheus::Gauge*> _connections_gauge;
-    std::map<const std::string, prometheus::Histogram*> _process_msg_timer;
-    std::map<const std::string, prometheus::Counter*> _push_msg_counter;
-    std::vector<prometheus::Counter*> _tx_validations;
-    prometheus::Counter* _bad_peer_counter;
-    prometheus::Counter* _misbehave_counter;
-    prometheus::Gauge* _known_peers_gauge;
-    prometheus::Gauge* _banned_gauge;
-    prometheus::Summary* _send_msg_timer;
-    std::map<const std::string, prometheus::Gauge*> _permission_gauge;
-
-public:
-    ~PeerMetricsImpl(){};
-    explicit PeerMetricsImpl(const std::string& chain, prometheus::Registry& registry);
-    void ProcessMsgType(const std::string& msg_type, long amt) override;
-    void IncTxValidationResult(int state) override;
-    void IncDiscourage() override;
-    void IncMisbehaveAmount(int amt) override;
-    void ConnectionType(int type, uint amt) override;
-    void Permission(const std::string& permission, uint amt) override;
-    void Known(size_t amt) override;
-    void SendMessageTime(long amt) override;
-    void PushMsgType(const std::string& msg_type) override;
-    void Banned(unsigned long amt) override;
-};
-
-class MemPoolMetrics
-{
-public:
-    virtual ~MemPoolMetrics() {}
-    static std::unique_ptr<MemPoolMetrics> make(const std::string& chain, prometheus::Registry& registry, bool noop);
-    virtual void AcceptTime(long amt){};
-    virtual void Transactions(MemPoolType type, long amt){};
-    virtual void Incoming(size_t in, size_t out, unsigned int byte_size, int64_t amt){};
-    virtual void Removed(size_t reason){};
-    virtual void Orphans(size_t map, size_t outpoint){};
-
-    virtual std::optional<double> GetRemoved(size_t reason) { return std::nullopt; };
-};
-class MemPoolMetricsImpl : virtual public MemPoolMetrics, Metrics
-{
-protected:
-    prometheus::Histogram* _accept_pool_timer;
-    std::vector<prometheus::Gauge*> _mempool_gauge;
-    prometheus::Counter* _incoming_counter;
-    prometheus::Counter* _vin_incoming_counter;
-    prometheus::Counter* _vout_incoming_counter;
-    prometheus::Counter* _incoming_size_counter;
-    prometheus::Counter* _incoming_amt_counter;
-    std::vector<prometheus::Counter*> _removed_counter;
-    prometheus::Counter* _removed_unknown_counter;
-    prometheus::Gauge* _orphan_size_gauge;
-    prometheus::Gauge* _orphan_outpoint_gauge;
-
-public:
-    explicit MemPoolMetricsImpl(const std::string& chain, prometheus::Registry& registry);
-    ~MemPoolMetricsImpl(){};
-    void AcceptTime(long amt) override;
-    void Transactions(MemPoolType type, long amt) override;
-    void Incoming(size_t in, size_t out, unsigned int byte_size, int64_t amt) override;
-    void Removed(size_t reason) override;
-    void Orphans(size_t map, size_t outpoint) override;
-
-    std::optional<double> GetRemoved(size_t reason) override;
-};
 
 
 } // namespace metrics
